Pull out logic to parse PNC dates into a separate file
