+++ src/use-cases/enrichHearingOutcome/enrichFunctions/enrichOffenceResultsPostPncEnrichment/populateSourceOrganisation.ts
@@ -1,6 +1,6 @@
 import type { AnnotatedHearingOutcome, Result } from "src/types/AnnotatedHearingOutcome"
-import populateOrganisationUnitFields from "src/use-cases/populateOrganisationUnitFields"
 import { lookupOrganisationUnitByThirdLevelPsaCode } from "src/use-cases/dataLookup"
+import populateOrganisationUnitFields from "src/use-cases/populateOrganisationUnitFields"
 
 const populateSourceOrganisation = (result: Result, hearingOutcome: AnnotatedHearingOutcome) => {
   const { CourtHearingLocation, CourtHouseCode } = hearingOutcome.AnnotatedHearingOutcome.HearingOutcome.Hearing
