+++ tests/helpers/generateMockPncQueryResultFromAho.ts
@@ -1,4 +1,5 @@
 import { XMLParser } from "fast-xml-parser"
+import parsePncDate from "src/lib/parsePncDate"
 import type { PncOffence, PncQueryResult } from "src/types/PncQueryResult"
 
 type OffenceDates = {
@@ -51,20 +52,6 @@ type ParsedAHO = {
       }
     }
   }
-}
-
-const parsePncDate = (pncDate: string): Date | null => {
-  if (!pncDate || pncDate === "") {
-    return null
-  }
-  const parsedDate = pncDate.match(/(\d{2})(\d{2})(\d{4})/)
-  if (!parsedDate) {
-    return null
-  }
-  const day = parseInt(parsedDate[1], 10)
-  const month = parseInt(parsedDate[2], 10)
-  const year = parseInt(parsedDate[3], 10)
-  return new Date(Date.UTC(year, month - 1, day))
 }
 
 const extractDates = (offence: AhoPncOffence): OffenceDates => {
